# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Booking Platform API
  description: "Event/ticket booking platform — users, events, venues, bookings, seat\
    \ holds, stubbed payment, and admin moderation."
  contact:
    name: Booking Platform
  license:
    name: Proprietary
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Local (dev)
security:
- bearerAuth: []
paths:
  /api/venues:
    get:
      tags:
      - venue-controller
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VenueResponse"
    post:
      tags:
      - venue-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponse"
  /api/events:
    get:
      tags:
      - event-controller
      operationId: listPublished
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponse"
    post:
      tags:
      - event-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/events/{id}/publish:
    post:
      tags:
      - event-controller
      operationId: publish
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/events/{id}/cancel:
    post:
      tags:
      - event-controller
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/bookings:
    get:
      tags:
      - booking-controller
      operationId: mine
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
    post:
      tags:
      - booking-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/bookings/{id}/pay:
    post:
      tags:
      - booking-controller
      operationId: pay
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/bookings/{id}/cancel:
    post:
      tags:
      - booking-controller
      operationId: cancel_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /api/admin/users/{id}/enable:
    post:
      tags:
      - admin-controller
      operationId: enableUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/admin/users/{id}/disable:
    post:
      tags:
      - admin-controller
      operationId: disableUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/admin/events/{id}/cancel:
    post:
      tags:
      - admin-controller
      operationId: cancelEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/venues/{id}:
    get:
      tags:
      - venue-controller
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponse"
  /api/users/me:
    get:
      tags:
      - user-controller
      operationId: me
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/ping:
    get:
      tags:
      - ping-controller
      operationId: ping
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/events/{id}:
    get:
      tags:
      - event-controller
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/events/mine:
    get:
      tags:
      - event-controller
      operationId: mine_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponse"
  /api/bookings/{id}:
    get:
      tags:
      - booking-controller
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/admin/users:
    get:
      tags:
      - admin-controller
      operationId: listUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
components:
  schemas:
    VenueRequest:
      required:
      - address
      - city
      - name
      type: object
      properties:
        name:
          maxLength: 150
          minLength: 0
          type: string
        address:
          maxLength: 250
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 0
          type: string
        capacity:
          minimum: 1
          type: integer
          format: int32
    VenueResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        city:
          type: string
        capacity:
          type: integer
          format: int32
    CreateEventRequest:
      required:
      - endsAt
      - startsAt
      - ticketTypes
      - title
      - venueId
      type: object
      properties:
        title:
          maxLength: 200
          minLength: 0
          type: string
        description:
          maxLength: 2000
          minLength: 0
          type: string
        venueId:
          type: integer
          format: int64
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        ticketTypes:
          type: array
          items:
            $ref: "#/components/schemas/TicketTypeRequest"
    TicketTypeRequest:
      required:
      - currency
      - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 0
          type: string
        priceCents:
          type: integer
          format: int64
        currency:
          pattern: "[A-Z]{3}"
          type: string
        quantityTotal:
          minimum: 1
          type: integer
          format: int32
    EventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        venueId:
          type: integer
          format: int64
        venueName:
          type: string
        organizerId:
          type: integer
          format: int64
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - CANCELLED
        ticketTypes:
          type: array
          items:
            $ref: "#/components/schemas/TicketTypeResponse"
    TicketTypeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        priceCents:
          type: integer
          format: int64
        currency:
          type: string
        quantityTotal:
          type: integer
          format: int32
        quantityAvailable:
          type: integer
          format: int32
    CreateBookingRequest:
      required:
      - eventId
      - items
      type: object
      properties:
        eventId:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/Line"
    Line:
      required:
      - ticketTypeId
      type: object
      properties:
        ticketTypeId:
          type: integer
          format: int64
        quantity:
          minimum: 1
          type: integer
          format: int32
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        attendeeId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - PAID
          - EXPIRED
          - CANCELLED
        totalCents:
          type: integer
          format: int64
        currency:
          type: string
        paymentReference:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    Item:
      type: object
      properties:
        ticketTypeId:
          type: integer
          format: int64
        ticketTypeName:
          type: string
        quantity:
          type: integer
          format: int32
        unitPriceCents:
          type: integer
          format: int64
    RegisterRequest:
      required:
      - displayName
      - email
      - password
      - role
      type: object
      properties:
        email:
          type: string
        password:
          maxLength: 100
          minLength: 8
          type: string
        displayName:
          maxLength: 100
          minLength: 0
          type: string
        role:
          type: string
          enum:
          - ATTENDEE
          - ORGANIZER
          - ADMIN
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - ATTENDEE
          - ORGANIZER
          - ADMIN
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
  securitySchemes:
    bearerAuth:
      type: http
      description: "JWT obtained from POST /api/auth/login. Send as: Authorization:\
        \ Bearer <token>"
      scheme: bearer
      bearerFormat: JWT